Capable of JPEG File Dump (as a degenerate version)
The JPEG files that the satellite created turned out bluish (the Red and Blue values were swapped, since JPEG reads the BMP file from end to start which is exactly what the std loader did [with some additional features], so I made it so our BMP file's planes were at the order of RGB instead of BGR).

In addition, Added all necessities for the image dump functionality.

Note: In this version of the code we could not send an image due to failures in the DataBase, Although the functions required for sending an image (chunkField_imageDump, JPEG_compressor, etc.) on their own are capable of downloading an image.
a degenerate main code not included.
(can be found at the "fileserver\LAVIAN\ImageDataBase & Camera\13.8.19\1 - fixed the jpeg (switched R and B at the BMP creation), also sent jpeg to ground" alongwith images recived at the ground station)