Cam Module Automatic Thumbnail Creation Update & Change to Error Logging
I made sure that unwanted results were not obtained after automatic thumbnail creation. If there is an RAW file for the image, I will not delete it after compression, which will happen if I created this file automatically.

Now when the Camera Module saves an error in the log, I will save the command's ID instead of select parameters from the command (since the commands parameters is saved in the ground...).